Removed mobile hiding of panels
Also added a 'first' attribute for panels. This will put a panel at the beginning visually (by placing flex order: -1), even though it is second in html.

This is for use cases where we semantically want to start with the  main content, and THEN then menu. While in the browser we want to see the menu first, and THEN the main content.

Always keep the right order of your html elements tidy and according to accessibility guidelines.